requirements.txt
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
reportlab==4.0.8